+++ docs/releases.rst
@@ -6,6 +6,8 @@ Version 1.5, In Development
 
 * ``TAPTestRunner`` has a ``set_combined`` method to collect all
   results in a single output file.
+* The ``nosetests`` plugin has an optional ``--tap-combined`` flag to
+  collect all results in a single output file.
 * ``TAPTestRunner`` has a ``set_format`` method to specify line format.
 * The ``nosetests`` plugin has an optional ``--tap-format`` flag to specify
   line format.